* Added `Yuni::Thread::ID`, to know the id of the current calling thread  * Fixed return value of `Yuni::ProcessID`
git-svn-id: file:///home/michael/svn/yuni/trunk@1941 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/system/process.cpp
@@ -1,5 +1,4 @@
 
-#include "../../yuni.h"
 #include "process.h"
 #ifndef YUNI_OS_WINDOWS
 #	ifndef YUNI_OS_HAIKU
@@ -22,9 +21,9 @@
 namespace Yuni
 {
 
-	unsigned int ProcessID()
+	uint64 ProcessID()
 	{
-		return (unsigned int) YUNI_OS_GETPID();
+		return (uint64) YUNI_OS_GETPID();
 	}
